requirements.txt
uvicorn
peewee
mariadb
fastapi
sqlalchemy
pymysql
pydactic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
redis